requirements.txt
mysql-connector-python~=9.1.0
app~=0.0.1
SQLAlchemy~=2.0.36
flask~=3.1.0
flasgger~=0.9.7.1
flask-restx~=1.3.0